// src/traveler/interfaces/traveler-info.interface.ts
export interface TravelerInfoInterface {
  organizationId: string;
  eventId: any;
  userId: string;
  tvChannel: string;
  position: string;
  outboundOriginCity: string;
  outboundFlightNumber: string;
  outboundArrivalTime: string;
  returnOriginCity: string;
  returnFlightNumber: string;
  returnArrivalTime: string;
  dietaryRestrictions: string;
}